// src/api/generated/models/CustomerResponse.ts
export interface CustomerResponse {
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    customerId?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    customerName?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    customerCode?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    customerType?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    industry?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    companySize?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    website?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    address?: string;
    /**
     * 
     * @type {number}
     * @memberof CustomerResponse
     */
    annualRevenue?: number;
    /**
     * 
     * @type {number}
     * @memberof CustomerResponse
     */
    employeeCount?: number;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    assignedUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    assignedUserName?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    branchId?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    branchName?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof CustomerResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CustomerResponse
     */
    updatedDate?: Date;
}